requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
openai==1.54.0
python-dotenv==1.0.1
numpy==1.26.3
pandas==2.2.0
werkzeug==3.0.1
pydantic==2.5.0
PyPDF2==3.0.1
scikit-learn==1.3.0
redis==5.0.1
aiofiles==23.2.0
bcrypt==4.0.1
PyJWT==2.8.0
cryptography==41.0.7
reportlab==4.0.7
aiohttp==3.9.1